requirements.txt
numpy
tensorflow
scikit-learn
seaborn
matplotlib
Pillow
tk
